+++ plugins/graphql-backend-module-catalog/src/entitiesLoadFn.ts
@@ -1,14 +1,19 @@
 import type { CatalogApi } from '@backstage/catalog-client';
 import { Entity } from '@backstage/catalog-model';
-import { NodeQuery } from '@frontside/hydraphql';
+import { getBearerTokenFromAuthorizationHeader } from '@backstage/plugin-auth-node';
+import { GraphQLContext, NodeQuery } from '@frontside/hydraphql';
 import { GraphQLError } from 'graphql';
+import type { Request } from 'node-fetch';
 import { CATALOG_SOURCE } from './constants';
 
 export const createCatalogLoader = (catalog: CatalogApi) => ({
   [CATALOG_SOURCE]: async (
     queries: readonly (NodeQuery | undefined)[],
+    context: GraphQLContext & { request?: Request }
   ): Promise<Array<Entity | GraphQLError>> => {
     // TODO: Support fields
+    const request = context.request;
+    const token = getBearerTokenFromAuthorizationHeader(request?.headers.get('authorization'));
     const entityRefs = queries.reduce(
       (refs, { ref } = {}, index) => (ref ? refs.set(index, ref) : refs),
       new Map<number, string>(),
@@ -16,7 +21,7 @@ export const createCatalogLoader = (catalog: CatalogApi) => ({
     const refEntries = [...entityRefs.entries()];
     const result = await catalog.getEntitiesByRefs({
       entityRefs: refEntries.map(([, ref]) => ref),
-    });
+    }, { token });
     const entities: (Entity | GraphQLError)[] = Array.from({
       length: queries.length,
     });