+++ plugins/graphql-backend-module-catalog/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @frontside/backstage-plugin-graphql-backend-module-catalog
 
+## 0.2.3
+
+### Patch Changes
+
+- e3d21a3: Corrects the request type used for extracting auth token for catalog requests.
+
+## 0.2.2
+
+### Patch Changes
+
+- 2124df0: Pass Backstage auth token to Catalog client requests
+
 ## 0.2.1
 
 ### Patch Changes